gigaslash mostly works

+++ src/main/java/com/favouritedragon/dragonenchantments/common/util/DragonUtils.java
@@ -182,11 +182,9 @@ public class DragonUtils {
 			entity.setPositionAndUpdate(d0, d1, d2);
 			return false;
 		} else {
-
 			if (entity instanceof EntityCreature) {
 				((EntityCreature) entity).getNavigator().clearPath();
 			}
-
 			return true;
 		}
 	}
@@ -225,10 +223,10 @@ public class DragonUtils {
 
 	//Vec3d from pitch and yaw of an entity
 	public static Vec3d getVectorForRotation(float pitch, float yaw) {
-		float f = MathHelper.cos(-yaw * 0.017453292F - (float)Math.PI);
-		float f1 = MathHelper.sin(-yaw * 0.017453292F - (float)Math.PI);
+		float f = MathHelper.cos(-yaw * 0.017453292F - (float) Math.PI);
+		float f1 = MathHelper.sin(-yaw * 0.017453292F - (float) Math.PI);
 		float f2 = -MathHelper.cos(-pitch * 0.017453292F);
 		float f3 = MathHelper.sin(-pitch * 0.017453292F);
-		return new Vec3d((double)(f1 * f2), (double)f3, (double)(f * f2));
+		return new Vec3d(f1 * f2, f3, f * f2);
 	}
 }
